test: simplify repro and drop duplicate stub
Consolidate on the single repro in optimize.rs, extract a small
append_fragment helper, and drop the duplicate stub in index/create.rs
that could not hit the bug (plan_compaction separates indexed from
unindexed fragments).
